fix: improve scripts

+++ docker/deploy.sh
@@ -5,21 +5,20 @@ if [ "$1" != "chrome" ] && [ "$1" != "firefox" ]; then
   exit 1
 fi
 
-npm run build && \
+if [ "$TRAVIS_BRANCH" == "master" ]; then
+  npm run build && \
 
-mkdir -p "docker/cybernaut-$1/cybernaut" && \
-cp -f "config-schema.json" "docker/cybernaut-$1/cybernaut/config-schema.json" && \
-cp -rf "dist" "docker/cybernaut-$1/cybernaut/dist" && \
-cp -f "package.json" "docker/cybernaut-$1/cybernaut/package.json" && \
+  mkdir -p "docker/cybernaut-$1/cybernaut" && \
+  cp -f "config-schema.json" "docker/cybernaut-$1/cybernaut/config-schema.json" && \
+  cp -rf "dist" "docker/cybernaut-$1/cybernaut/dist" && \
+  cp -f "package.json" "docker/cybernaut-$1/cybernaut/package.json" && \
 
-#if [ "$TRAVIS_BRANCH" == "master" ]; then
-cp -f "docker/cybernaut-xvfb" "docker/cybernaut-$1/cybernaut-xvfb" && \
+  cp -f "docker/cybernaut-xvfb" "docker/cybernaut-$1/cybernaut-xvfb" && \
 
-docker build -t clebert/cybernaut-$1 "docker/cybernaut-$1"
-#docker build -t clebert/cybernaut-$1 "docker/cybernaut-$1" && \
-#docker login -u="$DOCKER_USERNAME" -p="$DOCKER_PASSWORD" && \
-#docker push clebert/cybernaut-$1
+  docker build -t clebert/cybernaut-$1 "docker/cybernaut-$1" && \
+  docker login -u="$DOCKER_USERNAME" -p="$DOCKER_PASSWORD" && \
+  docker push clebert/cybernaut-$1
 
-rm -rf "docker/cybernaut-$1/cybernaut"
-rm -f "docker/cybernaut-$1/cybernaut-xvfb"
-#fi
+  rm -rf "docker/cybernaut-$1/cybernaut"
+  rm -f "docker/cybernaut-$1/cybernaut-xvfb"
+fi
